Moving some code to a new trait
