+++ src/ORM/Association/ExternalAssociationTrait.php
@@ -15,7 +15,7 @@
 namespace Cake\ORM\Association;
 
 use Cake\Database\Expression\IdentifierExpression;
-use Cake\Database\Expression\TupleComparison;
+use Cake\ORM\Association\SelectableAssociationTrait;
 use Cake\ORM\Query;
 use Cake\Utility\Inflector;
 
@@ -24,6 +24,8 @@ use Cake\Utility\Inflector;
  * a extra query.
  */
 trait ExternalAssociationTrait {
+
+	use SelectableAssociationTrait;
 
 /**
  * Order in which target records should be returned
@@ -72,18 +74,6 @@ trait ExternalAssociationTrait {
 			$this->_sort = $sort;
 		}
 		return $this->_sort;
-	}
-
-/**
- * Returns true if the eager loading process will require a set of parent table's
- * primary keys in order to use them as a filter in the finder query.
- *
- * @param array $options
- * @return boolean true if a list of keys will be required
- */
-	public function requiresKeys($options = []) {
-		$strategy = isset($options['strategy']) ? $options['strategy'] : $this->strategy();
-		return $strategy !== parent::STRATEGY_SUBQUERY;
 	}
 
 /**
@@ -197,127 +187,6 @@ trait ExternalAssociationTrait {
 			}
 			return $row;
 		};
-	}
-
-/**
- * Auxiliary function to construct a new Query object to return all the records
- * in the target table that are associated to those specified in $options from
- * the source table
- *
- * @param array $options options accepted by eagerLoader()
- * @return \Cake\ORM\Query
- * @throws \InvalidArgumentException When a key is required for associations but not selected.
- */
-	protected function _buildQuery($options) {
-		$target = $this->target();
-		$alias = $target->alias();
-		$key = $this->_linkField($options);
-
-		$filter = $options['keys'];
-		if ($options['strategy'] === parent::STRATEGY_SUBQUERY) {
-			$filter = $this->_buildSubquery($options['query'], $key);
-		}
-
-		$fetchQuery = $this
-			->find('all')
-			->where($options['conditions'])
-			->eagerLoaded(true)
-			->hydrate($options['query']->hydrate());
-		$fetchQuery = $this->_addFilteringCondition($fetchQuery, $key, $filter);
-
-		if (!empty($options['fields'])) {
-			$fields = $fetchQuery->aliasFields($options['fields'], $alias);
-			if (!in_array($key, $fields)) {
-				throw new \InvalidArgumentException(
-					sprintf('You are required to select the "%s" field', $key)
-				);
-			}
-			$fetchQuery->select($fields);
-		}
-
-		if (!empty($options['sort'])) {
-			$fetchQuery->order($options['sort']);
-		}
-
-		if (!empty($options['contain'])) {
-			$fetchQuery->contain($options['contain']);
-		}
-
-		if (!empty($options['queryBuilder'])) {
-			$options['queryBuilder']($fetchQuery);
-		}
-
-		return $fetchQuery;
-	}
-
-/**
- * Appends any conditions required to load the relevant set of records in the
- * target table query given a filter key and some filtering values.
- *
- * @param \Cake\ORM\Query taget table's query
- * @param string $key the fields that should be used for filtering
- * @param mixed $filter the value that should be used to match for $key
- * @return \Cake\ORM\Query
- */
-	protected function _addFilteringCondition($query, $key, $filter) {
-		if (is_array($key)) {
-			$types = [];
-			$defaults = $query->defaultTypes();
-			foreach ($key as $k) {
-				if (isset($defaults[$k])) {
-					$types[] = $defaults[$k];
-				}
-			}
-			return $query->andWhere(new TupleComparison($key, $filter, $types, 'IN'));
-		}
-
-		return $query->andWhere([$key . ' IN' => $filter]);
-	}
-
-/**
- * Generates a string used as a table field that contains the values upon
- * which the filter should be applied
- *
- * @param array $options
- * @return string
- */
-	protected function _linkField($options) {
-		$links = [];
-		$name = $this->name();
-
-		foreach ((array)$options['foreignKey'] as $key) {
-			$links[] = sprintf('%s.%s', $name, $key);
-		}
-
-		if (count($links) === 1) {
-			return $links[0];
-		}
-
-		return $links;
-	}
-
-/**
- * Builds a query to be used as a condition for filtering records in in the
- * target table, it is constructed by cloning the original query that was used
- * to load records in the source table.
- *
- * @param \Cake\ORM\Query $query the original query used to load source records
- * @param string $foreignKey the field to be selected in the query
- * @return \Cake\ORM\Query
- */
-	protected function _buildSubquery($query, $foreignKey) {
-		$filterQuery = clone $query;
-		$filterQuery->limit(null);
-		$filterQuery->contain([], true);
-		$joins = $filterQuery->join();
-		foreach ($joins as $i => $join) {
-			if (strtolower($join['type']) !== 'inner') {
-				unset($joins[$i]);
-			}
-		}
-		$filterQuery->join($joins, [], true);
-		$fields = $query->aliasFields((array)$query->repository()->primaryKey());
-		return $filterQuery->select($fields, true);
 	}
 
 /**
